requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decouple==0.0.7
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jwcrypto==0.8
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PySocks==1.7.1
python-decouple==3.3
python-dotenv==0.14.0
python-firebase==1.2
python-jwt==3.3.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
sseclient==0.0.27
tweepy==3.9.0
urllib3==1.25.10
